+++ plugin.d.ts
@@ -1,0 +1,3 @@
+declare module '@vueform/builder/plugin' {
+  export default function vueformPluginBuilder(): any;
+}