+++ vite.d.ts
@@ -1,0 +1,3 @@
+declare module '@vueform/builder/vite' {
+  export default function vitePluginVueform(): any;
+}